+++ test/lightning_web/live/project_live_test.exs
@@ -191,7 +191,7 @@ defmodule LightningWeb.ProjectLiveTest do
         |> render_submit()
         |> follow_redirect(conn, ~p"/projects")
 
-      assert html =~ "Project scheduled for deletion"
+      assert html =~ "Project scheduled for deletion later"
     end
 
     test "project members can export a project", %{conn: conn, project: project} do
@@ -348,7 +348,8 @@ defmodule LightningWeb.ProjectLiveTest do
         |> render_submit()
         |> follow_redirect(conn, Routes.project_index_path(conn, :index))
 
-      assert html =~ "Project deleted"
+      assert html =~
+               "Project deletion initiated. This could take upto 10 minutes, depending on the size of your data"
 
       refute index_live |> element("project-#{project.id}") |> has_element?()
     end